fix: address PR concerns
